MAINT: Separate export config resolution
